Separate getValue to getValue and getValueByPath.
