+++ src/ArrayHelper.php
@@ -183,14 +183,6 @@ class ArrayHelper
      * Retrieves the value of an array element or object property with the given key or property name.
      * If the key does not exist in the array or object, the default value will be returned instead.
      *
-     * The key may be specified in a dot format to retrieve the value of a sub-array or the property
-     * of an embedded object. In particular, if the key is `x.y.z`, then the returned value would
-     * be `$array['x']['y']['z']` or `$array->x->y->z` (if `$array` is an object). If `$array['x']`
-     * or `$array->x` is neither an array nor an object, the default value will be returned.
-     * Note that if the array already has an element `x.y.z`, then its value will be returned
-     * instead of going through the sub-arrays. So it is better to be done specifying an array of key names
-     * like `['x', 'y', 'z']`.
-     *
      * Below are some usage examples,
      *
      * ```php
@@ -236,26 +228,6 @@ class ArrayHelper
             return static::getRootValue($array, $lastKey, $default);
         }
 
-        if (is_array($array) && array_key_exists((string)$key, $array)) {
-            return $array[$key];
-        }
-
-        if (strpos($key, '.') !== false) {
-            foreach (explode('.', $key) as $part) {
-                if (is_array($array) && !array_key_exists($part, $array)) {
-                    return $default;
-                }
-
-                if (is_array($array) || is_object($array)) {
-                    $array = self::getValue($array, $part);
-                } else {
-                    return $default;
-                }
-            }
-
-            return $array;
-        }
-
         return static::getRootValue($array, $key, $default);
     }
 
@@ -283,6 +255,51 @@ class ArrayHelper
         }
 
         return $default;
+    }
+
+    /**
+     * Retrieves the value of an array element or object property with the given key or property name.
+     * If the key does not exist in the array or object, the default value will be returned instead.
+     *
+     * The key may be specified in a dot format to retrieve the value of a sub-array or the property
+     * of an embedded object. In particular, if the key is `x.y.z`, then the returned value would
+     * be `$array['x']['y']['z']` or `$array->x->y->z` (if `$array` is an object). If `$array['x']`
+     * or `$array->x` is neither an array nor an object, the default value will be returned.
+     * Note that if the array already has an element `x.y.z`, then its value will be returned
+     * instead of going through the sub-arrays. So it is better to be done specifying an array of key names
+     * like `['x', 'y', 'z']`.
+     *
+     * Below are some usage examples,
+     *
+     * ```php
+     * // working with array
+     * $username = \Yiisoft\Arrays\ArrayHelper::getValue($_POST, 'username');
+     * // working with object
+     * $username = \Yiisoft\Arrays\ArrayHelper::getValue($user, 'username');
+     * // working with anonymous function
+     * $fullName = \Yiisoft\Arrays\ArrayHelper::getValue($user, function ($user, $defaultValue) {
+     *     return $user->firstName . ' ' . $user->lastName;
+     * });
+     * // using dot format to retrieve the property of embedded object
+     * $street = \Yiisoft\Arrays\ArrayHelper::getValue($users, 'address.street');
+     * // using an array of keys to retrieve the value
+     * $value = \Yiisoft\Arrays\ArrayHelper::getValue($versions, ['1.0', 'date']);
+     * ```
+     *
+     * @param array|object $array array or object to extract value from
+     * @param string|\Closure|array $path key name of the array element, an array of keys or property name
+     * of the object, or an anonymous function returning the value. The anonymous function signature should be:
+     * `function($array, $defaultValue)`.
+     * @param mixed $default the default value to be returned if the specified array key does not exist. Not used when
+     * getting value from an object.
+     * @return mixed the value of the element if found, default value otherwise
+     */
+    public static function getValueByPath($array, $path, $default = null)
+    {
+        if (is_string($path)) {
+            $path = explode('.', $path);
+        }
+        return static::getValue($array, $path, $default);
     }
 
     /**
